most of the mlcc parser
